// src/profile.enum.ts
export enum ProfileEnum {
  RELAX = "relax",
  ADVENTURER = "aventureiro",
  GUARANTEEED = "garantido",
  INTELECTUAL = "intelectual",
  DYNAMIC = "dinamico",
  JUST_LIVE_ONCE = "so-se-vive-uma-vez",
  SPIRITUAL = "espiritual",
  GASTRONOMIC = "gastronomico",
  ROUTINE_FAN = "fa-da-rotina",
  ALTERNATIVE = "alternativo",
  INSACIABLE = "insaciavel",
}